pyproject.toml
[tool.poetry]
name = "crawler"
version = "0.1.0"
description = ""
authors = ["AryeH"]

[tool.poetry.dependencies]
python = "^3.9"
celery = "5.2.2"
redis = "^4.2.0"
flower = {git = "https://github.com/mher/flower.git"}
pytest-monitor = "^1.6.3"
celery-singleton = "^0.3.1"
celery-redbeat = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = {extras = ["toml"], version = "^6.3.2"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
